fixed issue w/ emtpy first elements in Arrays
