+++ src/engine/render/Leaf.ts
@@ -1,7 +1,6 @@
 import { Observable } from 'rxjs';
 import { distinctUntilChanged, map } from 'rxjs/operators';
 import { ICompiledComponent } from '.';
-import { log } from '../../helpers/logPipe';
 import { ILeafComponent } from '../component/Leaf';
 
 export interface ITextRenderElement {
@@ -15,7 +14,6 @@ export function renderLeaf(
     component: ILeafComponent,
 ): Observable<ICompiledComponent> {
     return component.render$.pipe(
-        log('RENDER LEAF'),
         distinctUntilChanged(),
         map((data) => {
             const text = data != null ? data.toString() : '';
