+++ src/helpers/destroyer.ts
@@ -1,6 +1,7 @@
 import { Observable, ReplaySubject } from 'rxjs';
 
-export function destroyer(): [() => void, Observable<void>] {
+// returns a pair [callback, subject] that help with handling destruction
+export function createDestroyer(): [() => void, Observable<void>] {
     const destroy$ = new ReplaySubject<void>(1);
     const destroy = () => {
         destroy$.next(void 0);
